Divide queries file into new entities files
