+++ index.js
@@ -1,6 +1,10 @@
 const express = require('express')
 const bodyParser = require('body-parser')
-const db = require('./queries')
+const queryUsers = require('./queriesusers')
+const queryBooks = require('./querybooks')
+const queryCategories = require('./querycategories')
+const queryBookChecks = require('./querybookchecks')
+const queryReports = require('./queryreports')
 const app = express()
 const port = process.env.API_PORT
 
@@ -19,62 +23,62 @@ app.listen(port, () => {
     console.log(`App running on port ${port}.`)
   })
   //API of users
-  app.get('/users', db.getUsersAsync)
-  //app.get('/users', db.getUsers)
-  //app.get('/users/:id', db.getUserByID)
-  app.get('/users/:id', db.getUserByIDAsync)
-  //app.post('/users', db.createUser)
-  app.post('/users', db.createUserAsync)
-  //app.put('/users/:id', db.updateUser)
-  app.put('/users/:id', db.updateUserAsync)
-  //app.delete('/users/:id', db.deleteUser)
-  app.delete('/users/:id', db.deleteUserAsync)
+  app.get('/users', queryUsers.getUsersAsync)
+  //app.get('/users', queryUsers.getUsers)
+  //app.get('/users/:id', queryUsers.getUserByID)
+  app.get('/users/:id', queryUsers.getUserByIDAsync)
+  //app.post('/users', queryUsers.createUser)
+  app.post('/users', queryUsers.createUserAsync)
+  //app.put('/users/:id', queryUsers.updateUser)
+  app.put('/users/:id', queryUsers.updateUserAsync)
+  //app.delete('/users/:id', queryUsers.deleteUser)
+  app.delete('/users/:id', queryUsers.deleteUserAsync)
   
   //API of books
-  // app.get('/books', db.getBooks)
-  app.get('/books', db.getBooksAsync)
-  //app.get('/books/:id', db.getBookByID)
-  app.get('/books/:id', db.getBookByIDAsync)
-  //app.post('/books', db.createBook)
-  app.post('/books', db.createBookAsync)
-  //app.put('/books/:id', db.updateBook)
-  app.put('/books/:id', db.updateBookAsync)
-  //app.delete('/books/:id', db.deleteBook)
-  app.delete('/books/:id', db.deleteBookAsync)
+  // app.get('/books', queryBooks.getBooks)
+  app.get('/books', queryBooks.getBooksAsync)
+  //app.get('/books/:id', queryBooks.getBookByID)
+  app.get('/books/:id', queryBooks.getBookByIDAsync)
+  //app.post('/books', queryBooks.createBook)
+  app.post('/books', queryBooks.createBookAsync)
+  //app.put('/books/:id', queryBooks.updateBook)
+  app.put('/books/:id', queryBooks.updateBookAsync)
+  //app.delete('/books/:id', queryBooks.deleteBook)
+  app.delete('/books/:id', queryBooks.deleteBookAsync)
 
   //API of categories
-  //app.get('/categories', db.getCategories)
-  app.get('/categories', db.getCategoriesAsync)
-  //app.get('/categories/:id', db.getCategoryByID)
-  app.get('/categories/:id', db.getCategoryByIDAsync)
-  //app.post('/categories', db.createCategory)
-  app.post('/categories', db.createCategoryAsync)
-  //app.put('/categories/:id', db.updateCategory)
-  app.put('/categories/:id', db.updateCategoryAsync)
-  //app.delete('/categories/:id', db.deleteCategory)
-  app.delete('/categories/:id', db.deleteCategoryAsync)
+  //app.get('/categories', queryCategories.getCategories)
+  app.get('/categories', queryCategories.getCategoriesAsync)
+  //app.get('/categories/:id', queryCategories.getCategoryByID)
+  app.get('/categories/:id', queryCategories.getCategoryByIDAsync)
+  //app.post('/categories', queryCategories.createCategory)
+  app.post('/categories', queryCategories.createCategoryAsync)
+  //app.put('/categories/:id', queryCategories.updateCategory)
+  app.put('/categories/:id', queryCategories.updateCategoryAsync)
+  //app.delete('/categories/:id', queryCategories.deleteCategory)
+  app.delete('/categories/:id', queryCategories.deleteCategoryAsync)
 
   //API of bookchecks
-  //app.get('/bookchecks', db.getBookChecks)
-  app.get('/bookchecks', db.getBookChecksAsync) 
-  //app.get('/bookchecks/:id', db.getBookChecksByID)
-  app.get('/bookchecks/:id', db.getBookcheckByIDAsync)
-  //app.post('/bookchecks', db.createBookCheck)
-  app.post('/bookchecks', db.createBookCheckAsync)
-  //app.put('/bookchecks/:id', db.updateBookCheck)
-  app.put('/bookchecks/:id', db.updateBookCheckAsync)
-  //app.delete('/bookchecks/:id', db.deleteBookCheck)
-  app.delete('/bookchecks/:id', db.deleteBookCheckAsync)
+  //app.get('/bookchecks', queryBookChecks.getBookChecks)
+  app.get('/bookchecks', queryBookChecks.getBookChecksAsync) 
+  //app.get('/bookchecks/:id', queryBookChecks.getBookChecksByID)
+  app.get('/bookchecks/:id', queryBookChecks.getBookcheckByIDAsync)
+  //app.post('/bookchecks', queryBookChecks.createBookCheck)
+  app.post('/bookchecks', queryBookChecks.createBookCheckAsync)
+  //app.put('/bookchecks/:id', queryBookChecks.updateBookCheck)
+  app.put('/bookchecks/:id', queryBookChecks.updateBookCheckAsync)
+  //app.delete('/bookchecks/:id', queryBookChecks.deleteBookCheck)
+  app.delete('/bookchecks/:id', queryBookChecks.deleteBookCheckAsync)
 
   //API of reports
-  //app.get('/reports/numberofusers', db.getNumberOfUsers)
-  app.get('/reports/numberofusers', db.getNumberOfUsersAsync)
-  //app.get('/reports/numberofusersbygender', db.getNumberOfUsersByGender)
-  app.get('/reports/numberofusersbygender', db.getNumberOfUsersByGenderAsync)
-  //app.get('/reports/numberofbooks', db.getNumberOfBooks)
-  app.get('/reports/numberofbooks', db.getNumberOfBooksAsync)
-  //app.get('/reports/numberofborrowedbooks', db.getNumberOfBorrowedBooks)
-  app.get('/reports/numberofborrowedbooks', db.getNumberOfBorrowedBooksAsync)
-  //app.get('/reports/numberofbooksbycategories', db.getNumberOfBooksByCategories)
-  app.get('/reports/numberofbooksbycategories', db.getNumberOfBooksByCategoriesAsync)
+  //app.get('/reports/numberofusers', queryReports.getNumberOfUsers)
+  app.get('/reports/numberofusers', queryReports.getNumberOfUsersAsync)
+  //app.get('/reports/numberofusersbygender', queryReports.getNumberOfUsersByGender)
+  app.get('/reports/numberofusersbygender', queryReports.getNumberOfUsersByGenderAsync)
+  //app.get('/reports/numberofbooks', queryReports.getNumberOfBooks)
+  app.get('/reports/numberofbooks', queryReports.getNumberOfBooksAsync)
+  //app.get('/reports/numberofborrowedbooks', queryReports.getNumberOfBorrowedBooks)
+  app.get('/reports/numberofborrowedbooks', queryReports.getNumberOfBorrowedBooksAsync)
+  //app.get('/reports/numberofbooksbycategories', queryReports.getNumberOfBooksByCategories)
+  app.get('/reports/numberofbooksbycategories', queryReports.getNumberOfBooksByCategoriesAsync)
